pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openrubricrl"
version = "0.1.0"
description = "Open-source pipeline that converts human-written rubrics into LLM-based reward functions for RL and RLHF training"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "OpenRubricRL Team", email = "team@openrubricrl.org"}
]
keywords = ["rlhf", "reinforcement-learning", "llm", "reward-function", "rubric"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "jsonschema>=4.17.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "click>=8.0.0",
    "aiohttp>=3.8.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
local = [
    "vllm>=0.2.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
]
metrics = [
    "nltk>=3.8.0",
    "rouge-score>=0.1.2",
    "sacrebleu>=2.3.0",
    "scikit-learn>=1.3.0",
]
all = [
    "openrubricrl[dev,local,metrics]"
]

[project.urls]
Homepage = "https://github.com/openrubricrl/openrubricrl"
Repository = "https://github.com/openrubricrl/openrubricrl"
Documentation = "https://openrubricrl.readthedocs.io"
"Bug Tracker" = "https://github.com/openrubricrl/openrubricrl/issues"

[project.scripts]
openrubricrl = "openrubricrl.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
asyncio_mode = "auto"